requirements.txt
langchain>=0.1.0
neo4j>=5.14.0
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.2
unstructured>=0.11.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6 